refactor: se ordenó la fomra en como estaba siendo consumido el recurso para crear cursos en moodle
- Se trasladó la función para crear curso al dominio especifico donde se utiliza.
- Se trasladó la función que genera el nombre corto del curso al controlador.
