+++ app/shared/services/moodle_service.py
@@ -90,70 +90,3 @@ class MoodleService:
             user_profile=user_profile,
             error=None,
         )
-
-    @staticmethod
-    def generate_shortname(institute: InstitutesEnum, fullname: str, group: str) -> str:
-        """
-        Genera el nombre corto oficial (Subject) siguiendo la lógica del proyecto.
-        Ejemplo: CIENCIAS + "Programación" + "G1" -> CIE-PRO-G1
-        """
-        inst_prefix = str(institute.value)[:3].upper()
-        words = fullname.split()
-
-        # Lógica de iniciales: toma la primera letra de las primeras 3 palabras
-        if len(words) >= 2:
-            course_initials = "".join([word[0] for word in words[:3]]).upper()
-        else:
-            course_initials = fullname[:3].upper()
-
-        group_suffix = str(group).upper() if group else "0"
-        return f"{inst_prefix}-{course_initials}-{group_suffix}"
-
-    @staticmethod
-    async def create_course(
-        institute: InstitutesEnum,
-        fullname: str,
-        teacher_name: str,
-        group_name: str,
-        category_id: int = 1,
-    ):
-        """
-        Automatiza la creación de un nuevo curso en Moodle.
-        Genera el shortname automáticamente antes de realizar la petición.
-        """
-        config = MOODLE_CONFIG[institute]
-        endpoint = config.moodle_url
-
-        # Generación de nomenclatura oficial
-        shortname_gen = MoodleService.generate_shortname(
-            institute, fullname, group_name
-        )
-        display_name = f"{fullname} - {group_name} ({teacher_name})"
-
-        params = {
-            "wstoken": config.moodle_token,
-            "wsfunction": "core_course_create_courses",
-            "moodlewsrestformat": "json",
-        }
-
-        data = {
-            "courses[0][fullname]": display_name,
-            "courses[0][shortname]": shortname_gen,
-            "courses[0][categoryid]": category_id,
-            "courses[0][idnumber]": group_name,
-            "courses[0][summary]": f"Docente: {teacher_name}",
-            "courses[0][format]": "topics",
-        }
-
-        result_response = await make_moodle_request(
-            url=endpoint,
-            params=params,
-            data=data,
-            institute=institute,
-        )
-
-        if not result_response["success"]:
-            return SimpleNamespace(course=None, error=result_response["error_message"])
-
-        # Moodle retorna una lista; extraemos el primer curso creado
-        return SimpleNamespace(course=result_response["data"][0], error=None)
